Update and improve metainfo (#205)
Improves the metainfo to better align with the Freedesktop.org specifications and Flathub standards, including adding proper release notes for this and all past Avogadro2 releases.

New commit used for `avogadroapp` points to a branch that simply adds [these changes to the metainfo](https://github.com/OpenChemistry/avogadroapp/commit/dcddde31975ec295970827949a935ab74dc09e24) onto the 1.100 release code. The build of the Avogadro binary itself remains identical and is the same as the original 1.100.0 release.